feat(stage-tamagotchi): initial dashboard ui
