--- apps/stage-tamagotchi/src/renderer/pages/dashboard/index.vue
@@ -1,5 +1,0 @@
-<template>
-  <div>
-    Dashboard
-  </div>
-</template>
